backend/requirements.txt
fastapi
uvicorn
langchain-google-genai
langchain-core
python-dotenv
pydantic
tenacity
requests
PyPDF2
fpdf2
python-multipart
